To be modular , remove mfg thing from stock module and put in mfg itselves
