+++ netforce_mfg/netforce_mfg/models/__init__.py
@@ -52,3 +52,13 @@ from . import production_plan
 from . import production_cost
 from . import production_period
 from . import production_period_cost
+from . import stock_picking
+from . import stock_move
+from . import stock_transform
+from . import stock_order
+from . import stock_barcode
+from . import stock_balance
+from . import report_stock_move
+from . import barcode_validate
+from . import barcode_validate_line
+
