+++ netforce_mfg/netforce_mfg/models/stock_barcode.py
@@ -1,15 +1,15 @@
 # Copyright (c) 2012-2015 Netforce Co. Ltd.
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
 # in the Software without restriction, including without limitation the rights
 # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be included in all
 # copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 # MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
@@ -21,20 +21,27 @@
 from netforce.model import Model, fields, get_model
 
 
-class BarcodeValidateLine(Model):
-    _name = "barcode.validate.line"
+class Barcode(Model):
+    _inhierit= "stock.barcode"
     _transient = True
     _fields = {
-        "wizard_id": fields.Many2One("barcode.validate", "Wizard", required=True, on_delete="cascade"),
-        "product_id": fields.Many2One("product", "Product", required=True),
-        "qty_planned": fields.Decimal("Planned Qty", required=True, readonly=True),
-        "qty_actual": fields.Decimal("Validated Qty"),
-        "uom_id": fields.Many2One("uom", "UoM", required=True),
-        "lot_id": fields.Many2One("stock.lot", "Lot / Serial Number"),
-        "container_to_id": fields.Many2One("stock.container", "Container"),
-        "location_from_id": fields.Many2One("stock.location", "From Location"),
-        "location_to_id": fields.Many2One("stock.location", "To Location"),
-        "related_id": fields.Reference([["sale.order", "Sales Order"], ["purchase.order", "Purchase Order"], ["job", "Service Order"], ["product.claim", "Claim Bill"], ["product.borrow", "Borrow Request"], ["stock.picking", "Picking"]], "Related To"),
+        "type": fields.Selection([["in", "Goods Receipt"], ["internal", "Goods Transfer"], ["out", "Goods Issue"]], "Transaction Type"),
+        "related_id": fields.Reference([["sale.order", "Sales Order"], ["purchase.order", "Purchase Order"], ["production.order", "Production Order"], ["stock.picking", "Picking"]], "Related To"),
     }
 
-BarcodeValidateLine.register()
+    def onchange_related(self, context={}):
+        data = context["data"]
+        type = data["type"]
+        val = data["related_id"][0]
+        relation, rel_id = val.split(",")
+        rel_id = int(rel_id)
+        if relation == "production.order":
+            rel = get_model("production.order").browse(rel_id)
+            if type == "out":
+                data["location_to_id"] = rel.production_location_id.id
+            elif type == "in":
+                data["location_from_id"] = rel.production_location_id.id
+        return data
+
+Barcode.register()
+
